FO-128 noe forbedring, prøvde å trekke ut actiontypes i egne filer.
