+++ src/frontend/src/ducks/types.ts
@@ -1,0 +1,15 @@
+
+export enum ActionType {
+    HENT_REGISTRERINGSTATUS,
+    HENTET_REGISTRERINGSTATUS,
+    HENT_REGISTRERINGSTATUS_FEILET,
+
+    HENT_INNLOGGINGSINFO,
+    HENTET_INNLOGGINGSINFO,
+    HENT_INNLOGGINGSINFO_FEILET,
+
+    AVGI_SVAR,
+    AVGA_SVAR,
+    AVGI_SVAR_FEILET,
+    AVGI_SVAR_INITIAL_STATE,
+}
